lib: Actually use provided size when creating BO
When reworking igt_create_fb_with_bo_size to use
igt_create_bo_with_dimensions, we mistakenly stopped acknowledging
non-zero size parameters.

To fix this, we move the core of the code to create_bo_for_fb and teach
it to use the GEM APIs when a size is passed.

igt_create_bo_with_dimensions ends up calling just create_bo_for_fb with
a zero size because now the later is more generic than the former.

Also, create_bo_for_fb now returns the handle of the BO that was
created, as there's no point anymore in having it be a parameter passed
by reference.

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>