tests: Add gem_exec_gttfill
This is an attempt to emulate the stressful nature of gem_cs_prefetch
without being as slow i.e. without trying to detect the GPU prefetch
issues and just emphasive the basic correctness in handling enough
batches to fill the GTT similar to gem_ringfill aims to fill a ring.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>