tests: Add kms_invalid_dotclock
Add a test that makes sure every modeset gets rejected by the kernel if
the requested dotclock is beyond the hardware capabilities.

For now we just test the preferred mode for each connector, should
perhaps test them all to be more sure everything is getting rejected.

We also skip the test on connectros that have a fixed mode as the kernel
will ignore most of the user timings. We should make the kernel more
strict I think, to at least check that the user gets roughly the refresh
rate they requeted.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>