igt_core: Search "." as final fallback for igt_fopen_data()
Some validation teams seem to run tests out of source directories that
have been nfs mounted or rsync'd to different locations on the target
machine.  This causes the igt_srcdir that the tests were built with to
be invalid on the machine the tests get run on.  Add the current
directory as a final fallback for data file searches.

Reference: https://bugs.freedesktop.org/show_bug.cgi?id=92248
Cc: Humberto Israel Perez Rodriguez <humberto.i.perez.rodriguez@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Marius Vlad <marius.c.vlad@intel.com>