lib: Start weaning off defunct intel_chipset.h
Several years ago we made the plan of only having one canonical source
for i915_pciids.h, the kernel and everyone importing their definitions
from that. For consistency, we style the intel_device_info after the
kernel, most notably using a generation mask and a per-codename bitfield.

This first step converts looking up the generation for a devid tree from
a massive if(devid)-chain to a (cached) table lookup.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>