igt_kms: Change PIPE_ANY behavior to mean unassigned, v2.
None of the tests requires that a output bound to PIPE_ANY is assigned,
so don't do it. Fix the display commit to iterate over crtc's instead
oof outputs to properly disable pipes without outputs.

This also means that output->valid is only set after connecting a
output to a pipe, so no longer depend on it in for_each_connected_output
and similar macros.

New code should use PIPE_NONE to avoid confusion, PIPE_ANY is a
deprecated alias, and will be removed.

Changes since v1:
- Split out the property changes.
- Split out the commit changes.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Ander Conselvan de Oliveira <conselvan2@gmail.com>