kms_atomic_transition: run correctly when hitting hw limits
On CHV with a 4K hdmi display it's possible to hit the wm upper
limit, 1 rgb primary plane + 2 full size rgb secondary planes
will hit a HW limit. To fix this try halving the horizontal width
until we're below HW limits. This requires a kernel patch to make
sure that the kernel considers actual width in wm calculations,
but without it we get FIFO underruns anyway.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>