igt/prime_vgem: Simplify inter-process flip checks
Avoid having both the child and parent do the same "did the flip"
happen check with each looking for the same event on the same fd. The
problem being that the child may fall asleep and by the time it wakes up
to do its check, the parent has already eaten the event. So leave the
checking that the flip does occur to the parent.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>