kms_cursor_legacy: Add cursor crc tests.
On skylake there's a add_all_affected_planes call that will make
atomic commit use the old cursor state. Add a test that first does
a page flip then cursor update to expose the issue. The test currently
fails on the vblank wait, but that's a common issue affecting all the
tests.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>