igt/gem_exec_nop: Relax assertion for parallel execution
In an ideal world, we should be able to execute on every engine in
parallel and the single limiting factor would be how fast the GPU can
execute. Due to the serialisation in execbuf, we would lockstep with
execution to the slowest engine and so would execute the same number of
cycles on each. However in CI, we are limited by how fast the driver is,
particularly under invasive debugging. This makes asserting that the
average time == max/nengine impossible, and reveals that the assertion is
impossible to meet under general condition. It's an impractical
regression test. Therefore we relax the assertion to only detect should
something critically fail. Worst case behaviour is presumed that each
ring runs sequentially, and so running N rings in parallel should take no
longer than running N rings serially. (Pathologically it can be even
slower if no batching on the rings occur).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>