lib: Add an iterator to generate primes
Primes are useful as input sources since they should not fall into any
patterns that may be optimised by the drivers.

An example of use this in a test driver:

	uint32_t seqno, inc, count;

	seqno = inc = count = 0;
	do {
		inc = igt_next_prime_number(inc);
		if (seqno + inc < seqno)
			break;
		seqno += inc;
		/* igt_assert_eq(test_inc(inc), seqno); */
		count++;
	} while (1);
	printf("count=%u, seqno=%u, last=%u\n", count, seqno, inc);

and prints "count=27878, seqno=4294845817, last=323381", or for simply
generating the set of the first N prime numbers, use

	for_each_prime_number(prime, 100)
		printf("%lu ", prime);

which prints

1 2 3 5 7 11 13 17 19 23 29 31 37 41 43 47 53 59 61 64 67 71 73 79 83 89
97 101 103 107 109 113 127 131 137 139 149 151 157 163 167 173 179 181
191 193 197 199 211 223 227 229 233 239 241 251 257 263 269 271 277 281
283 293 307 311 313 317 331 337 347 349 353 359 367 373 379 383 389 397
401 409 419 421 431 433 439 443 449 457 461 463 467 479 487 491 499 503
509 521

Note that 1 is included in the set of prime numbers for convenience
above. Mathematicians beware!

The set of primes is computed using the Sieve of Eratosthenes, which
basically just keeps a list of all multiples - any number not in list is
therefore prime. As a bitmask of all integers is kept, it can be quite
memory intensive for very large primes. A fallback to "trial division"
is available just in case, but for large primes that is much slower.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>