kms_atomic_transition: Respect limits for sprite planes on earlier platforms.
Usually there is a maximum amount of width/height that a sprite
plane can have, and on earlier platforms alpha sprite planes are
usually unsupported.

In case of SKL+ and VLV/CHV this is handled by only setting a maximum
width, but on earlier platforms there's a hardware limitation for
stride, width and height. Handle this by starting from cursor
dimensions, and then increase width to max, followed by height to max.

If the TEST_ONLY commit fails immediately with -EINVAL with cursor
width/height in ARGB format, assume that ARGB is unsupported and
fallback to XRGB. If that fails too fail the test.

This should allow tests to pass on all earlier platforms too.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>